Added the TOKEN_PRIVILEGES definition to the ndk.setypes library module.

+++ lib/ndk/datatypes.py
@@ -366,7 +366,7 @@ class LCID(DWORD): pass
 class LUID(pstruct.type):
     _fields_ = [
         (DWORD, 'LowPart'),
-        (DWORD, 'HighPart'),
+        (LONG, 'HighPart'),
     ]
 class PLUID(pointer(LUID)): pass
